refactor archive document caller context
