+++ src/app/routers/archive_document_downloads.py
@@ -4,12 +4,30 @@ from fastapi import APIRouter, Header, Path, Response
 
 from app.middleware.correlation import correlation_id_var
 from app.routers.archive_documents_common import (
-    archive_caller_headers,
+    ArchiveDocumentCallerHeaders,
     archive_error_response,
 )
 from app.services.gateway_service_provider import archive_document_service
 
 router = APIRouter(prefix="/api/v1/documents", tags=["Archived Documents"])
+
+
+async def _download_archived_document(
+    *,
+    document_id: str,
+    caller_headers: ArchiveDocumentCallerHeaders,
+) -> Response:
+    download = await archive_document_service().download_document(
+        document_id=document_id,
+        caller_headers=caller_headers.as_archive_context(),
+        correlation_id=correlation_id_var.get(),
+    )
+
+    return Response(
+        content=download.content,
+        media_type=download.media_type,
+        headers=download.headers,
+    )
 
 
 @router.get(
@@ -67,9 +85,9 @@ async def download_archived_document(
     booking_center_code: Annotated[str | None, Header(alias="X-Booking-Center-Code")] = None,
     role: Annotated[str | None, Header(alias="X-Role")] = None,
 ) -> Response:
-    download = await archive_document_service().download_document(
+    return await _download_archived_document(
         document_id=document_id,
-        caller_headers=archive_caller_headers(
+        caller_headers=ArchiveDocumentCallerHeaders(
             actor_id=actor_id,
             caller_application=caller_application,
             tenant_id=tenant_id,
@@ -77,11 +95,4 @@ async def download_archived_document(
             booking_center_code=booking_center_code,
             role=role,
         ),
-        correlation_id=correlation_id_var.get(),
-    )
-
-    return Response(
-        content=download.content,
-        media_type=download.media_type,
-        headers=download.headers,
     )
