+++ src/app/routers/archive_documents_common.py
@@ -1,3 +1,4 @@
+from dataclasses import dataclass
 from typing import Any
 
 from app.contracts.archive_documents import (
@@ -5,6 +6,26 @@ from app.contracts.archive_documents import (
     ArchivedDocumentErrorResponse,
 )
 from app.services.caller_context import caller_context_headers
+
+
+@dataclass(frozen=True)
+class ArchiveDocumentCallerHeaders:
+    actor_id: str | None
+    caller_application: str | None
+    tenant_id: str | None
+    region: str | None
+    booking_center_code: str | None
+    role: str | None
+
+    def as_archive_context(self) -> dict[str, str]:
+        return archive_caller_headers(
+            actor_id=self.actor_id,
+            caller_application=self.caller_application,
+            tenant_id=self.tenant_id,
+            region=self.region,
+            booking_center_code=self.booking_center_code,
+            role=self.role,
+        )
 
 
 def archive_error_response(
